requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
numpy==1.14.5
pandas==0.23.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
